requirements.txt
absl-py==0.7.0
altair==3.2.0
argh==0.26.2
astor==0.7.1
astropy==3.2.1
attrs==18.2.0
backcall==0.1.0
base58==1.0.3
bleach==3.1.0
blinker==1.4
boto3==1.9.249
botocore==1.12.249
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
enum34==1.1.6
future==0.18.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.3
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
notebook==5.7.4
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pathtools==0.1.2
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==4.1.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.20.0
pyparsing==2.3.1
pyrsistent==0.14.11
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.0.0
qtconsole==4.4.3
requests==2.21.0
retrying==1.3.3
s3transfer==0.2.1
scikit-image==0.15.0
scipy==1.3.0
Send2Trash==1.5.0
shapelets==0.2
six==1.12.0
streamlit==0.48.1
tb-nightly==1.14.0a20190301
tensorboard==1.13.0
tensorflow==2.0.0a0
tensorflow-estimator==1.13.0
tensorflow-gpu==2.0.0a0
tensorflow-probability==0.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
tfp-nightly==0.8.0.dev20190603
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
tzlocal==2.0.0
urllib3==1.24.2
validators==0.14.0
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
